update gmap, adaugat trances . de verificat url-urile

+++ rent-ejb/src/main/java/ro/agitman/model/Address.java
@@ -1,6 +1,7 @@
 package ro.agitman.model;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "rt_address")
@@ -16,8 +17,8 @@ public class Address extends AbstractModel {
     private String et;
     private String ap;
     private String cp;
-    private Double lat;
-    private Double lng;
+    private BigDecimal lat;
+    private BigDecimal lng;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -105,19 +106,19 @@ public class Address extends AbstractModel {
         this.cp = cp;
     }
 
-    public Double getLat() {
+    public BigDecimal getLat() {
         return lat;
     }
 
-    public void setLat(Double lat) {
+    public void setLat(BigDecimal lat) {
         this.lat = lat;
     }
 
-    public Double getLng() {
+    public BigDecimal getLng() {
         return lng;
     }
 
-    public void setLng(Double lng) {
+    public void setLng(BigDecimal lng) {
         this.lng = lng;
     }
 }
